separation of concerns
We separated all classess to its own.
Then we made own module 'core' for all the logic
