+++ src/core/DocumentCutter.ts
